pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "Nanodecoder"
version = "0.1.0"
description = "A foundation for experimenting with and extending Transformer architectures, focusing on decoder-only models like GPT"
readme = "README.md"
requires-python = ">=3.8"
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
dependencies = [
    "torch",
    "tiktoken",
    "matplotlib",
    "tensorflow",
    "tqdm",
    "numpy",
    "pandas",
    "psutil",
    "PyYAML",
    "wandb",
    "python-dotenv",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "black",
    "flake8",
    "mypy",
]

[tool.uv]
dev-dependencies = [
    "pytest",
    "black",
    "flake8", 
    "mypy",
]

[tool.black]
line-length = 88
target-version = ['py38']

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
